// src/modules/garmentsProduction/reports/embellishment/embellishment-pi-report/embellishment-pi-report-type.ts
export type EmbellishmentPIReportType = {
  ID: number;
  PI_DATE?: string | null;
  PARTY_ID?: number | null;
  ADDRESS?: string | null;
  PI_NO?: string | null;
  CREATED_BY?: string | null;
  UPDATED_BY?: string | null;
  CREATED_DATE?: string | null;
  UPDATED_DATE?: string | null;
  TERM_CONDITIONS?: string | null;
  DISCOUNT_PERCENT?: number | null;
  PARTY?: string | null;
  ID_1?: number | null;
  MASTER_ID?: number | null;
  WORK_ORDER_ID?: number | null;
  QTY?: number | null;
  DZN_PRICE?: number | null;
  AMOUNT?: number | null;
  EMB_WORK_ORDER_NO?: string | null;
  PRICE?: number | null;
  WORK_ORDER_NO?: string | null;
  WORK_ORDER_ID_1?: number | null;
  BUYER?: string | null;
  STYLE?: string | null;
  PO_NO?: string | null;
  PARTS?: string | null;
  OS_BUYER?: string | null;
  OS_STYLE?: string | null;
  OS_PO_NO?: string | null;
  PRINT_TYPE?: string | null;
  COMPANY_NAME?: string | null;
  COMPANY_ADDRESS?: string | null;
};